UE5 마이그레이션 #2
- 루멘, 나나이트 활성화
- 파티클 시스템 -> 나이아가라 시스템 컨버팅
  - 관련 플러그인 설치하여 컨버팅
- 코드 개선에 따른 변경사항 반영, 게임 로직 정상화

+++ Source/RoomEscapeFPS/Private/UI/InventoryPanel.cpp
@@ -60,7 +60,7 @@ void UInventoryPanel::UpdateNextInformation(ENextInformationType curType, ENextI
 {
 	TArray<FStringFormatArg> args;
 	args.Add(InCount);
-	NextInformationText->SetText(FText::FromString(FString::Format(*NextInformationStrArray[(uint8)curType], args)));
+	NextInformationText->SetText(FText::FromString(FString::Format(*NextInformationStrArray[static_cast<int32>(curType)], args)));
 	if (CurrentType != nextType)
 	{
 		CurrentType = nextType;
